+++ libs/common/constant_strings.h
@@ -122,6 +122,7 @@ ASTR(AA_seed);
 ASTR(ArnoldUsd);
 ASTR(ArnoldNodeGraph);
 ASTR(ArnoldOptions);
+ASTR(ArnoldMarkPrimsDirty);
 ASTR(binary);
 ASTR(BOOL);
 ASTR(BYTE);